some changes
I added minor changes and parameters that allow you to download the old version as the current WSA is problematic. Once the current wsa fixes the issues I will update the script and have it download the newest version. If you still want to install the latest version. Replace tempwsa=yes with another value at the beginning of the sh file, for example tempwsa=no